requirements.txt
Django==1.4
gondor==1.1.5
South==0.7.5
psycopg2==2.4.5
django-crispy-forms==1.1.3
django-rosetta==0.6.6
django-socialregistration==0.5.4
httplib2==0.7.4
oauth2==1.5.211
python-openid==2.2.5
simplejson==2.5.2
wsgiref==0.1.2
-e git+https://github.com/pythonforfacebook/facebook-sdk#egg=FacebookSDK
django-admin-tools==0.4.1
SQLAlchemy==0.7.8
django-celery==3.0.11
redis==2.7.2
django-picklefield==0.2.1
gunicorn==0.14.3
eventlet==0.9.16
celery==3.0.11
celerymon==1.0.3
